Before:

```
In [1]: from mimesis.providers import Address
   ...: f = Address._dd_to_dms

In [2]: %timeit f(57.12345, "lt")
1.18 µs ± 2.07 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

After:

```
In [2]: %timeit f(57.12345, "lt")
760 ns ± 2.61 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

# I have made things!

<!--
Hi, thanks for submitting a Pull Request. We appreciate it.

Please, fill in all the required information
to make our review and merging processes easier.

Cheers!
-->

## Checklist

<!-- Please check everything that applies: -->

- [x] I have read [contributing guidelines](https://github.com/lk-geimfari/mimesis/blob/master/CONTRIBUTING.rst)
- [x] I'm sure that I did not unrelated changes in this pull request
- [ ] I have created at least one test case for the changes I have made

<!-- Uncomment this section if this is your very first PR to this repository
- [ ] I have added myself to the [CONTRIBUTORS.rst](https://github.com/lk-geimfari/mimesis/blob/master/CONTRIBUTORS.rst)
-->

<!-- Uncomment this if documentation update required
- [ ] I have updated the documentation for the changes I have made
-->

- [x] I have added my changes to the [CHANGELOG.rst](https://github.com/lk-geimfari/mimesis/blob/master/CHANGELOG.rst)

## Related issues

<!--
Mark what issues this Pull Request closes or references.

Format is:
- Closes #issue-number
- Refs #issue-number

Example. Refs #0
Documentation: https://blog.github.com/2013-05-14-closing-issues-via-pull-requests/
-->

<!--
If you have any feedback, just write it here.

It can be whatever you want!
-->